function/food_recognition_utils.py
import numpy as np
from PIL import Image
import time
import streamlit as st
from data.food_database import FOOD_DATABASE, STANDARD_PORTIONS

# Simplified model loading function that doesn't require TensorFlow
@st.cache_resource
def load_model():
    """Simulate loading a model for food recognition"""
    # In a real implementation, you would load a pre-trained model
    # For this demo, we'll just return a placeholder
    return "food_recognition_model_placeholder"

# Simplified image preprocessing function
def preprocess_image(image, target_size=(224, 224)):
    """Preprocess the image for analysis"""
    if image.mode != "RGB":
        image = image.convert("RGB")
    image = image.resize(target_size)
    # In a real implementation, you would convert the image to a format suitable for the model
    # For this demo, we'll just return the resized image
    return image

# Function to predict food from image
def predict_food(model, image):
    """Predict food from image using the model"""
    # In a real implementation, you would use a fine-tuned model for food recognition
    # For this demo, we'll simulate predictions

    # Simulate model processing time
    time.sleep(2)

    # Simulate prediction (in a real app, you would use model.predict())
    # For demo purposes, we'll randomly select from our food database
    foods = list(FOOD_DATABASE.keys())
    predicted_food = np.random.choice(foods)

    # Simulate confidence scores
    confidence = np.random.uniform(0.7, 0.98)

    return predicted_food, confidence

# Function to estimate portion size
def estimate_portion_size(image, food_type):
    """Estimate portion size from image"""
    # In a real implementation, you would use 3D reconstruction or reference objects
    # For this demo, we'll simulate portion size estimation

    # Simulate processing time
    time.sleep(1)

    # Get standard portion for this food
    standard_portion = STANDARD_PORTIONS.get(food_type, 100)

    # Simulate a portion size variation (±30% from standard)
    variation = np.random.uniform(0.7, 1.3)
    estimated_portion = standard_portion * variation

    return round(estimated_portion, 1)

# Function to calculate nutritional information based on food and portion size
def calculate_nutrition(food, portion_size):
    """Calculate nutritional information based on food and portion size"""
    if food not in FOOD_DATABASE:
        return None

    # Get base nutritional values per 100g
    nutrition = FOOD_DATABASE[food]

    # Calculate nutrition based on portion size
    portion_factor = portion_size / 100.0

    # Calculate adjusted nutrition values
    adjusted_nutrition = {
        'carbs': round(nutrition['carbs'] * portion_factor, 1),
        'calories': round(nutrition['calories'] * portion_factor, 1),
        'protein': round(nutrition['protein'] * portion_factor, 1),
        'fat': round(nutrition['fat'] * portion_factor, 1),
        'fiber': round(nutrition['fiber'] * portion_factor, 1),
        'gi': nutrition['gi']
    }

    return adjusted_nutrition
