requirements.txt
dill==0.3.9
matplotlib==3.9.2
numpy==2.0.0
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
scikit-learn==1.5.2
seaborn==0.13.2
shap==0.46.0
streamlit==1.40.1
toml==0.10.2
opencv-python==4.8.0.76
